+++ src/Oro/Bundle/EntityBundle/Controller/DictionaryController.php
@@ -2,67 +2,50 @@
 
 namespace Oro\Bundle\EntityBundle\Controller;
 
-use Oro\Bundle\EntityBundle\Entity\Manager\DictionaryApiEntityManager;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Oro\Bundle\EntityBundle\Provider\DictionaryEntityDataProvider;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * Entities controller.
+ * The controller to get dictionary values.
  */
 #[Route(path: '/dictionary')]
-class DictionaryController extends AbstractController
+class DictionaryController
 {
+    private DictionaryEntityDataProvider $dictionaryEntityDataProvider;
+
+    public function __construct(DictionaryEntityDataProvider $dictionaryEntityDataProvider)
+    {
+        $this->dictionaryEntityDataProvider = $dictionaryEntityDataProvider;
+    }
+
     /**
-     * Get dictionary values by search query
-     *
-     * @param string $dictionary - Class Name Entity that was configured as Dictionary
-     *
-     *
-     * @return JsonResponse
+     * Gets dictionary values by search query.
      */
     #[Route(path: '/{dictionary}/search', name: 'oro_dictionary_search')]
-    public function searchAction($dictionary)
+    public function searchAction(string $dictionary, Request $request): JsonResponse
     {
-        $searchQuery = $this->container->get('request_stack')->getCurrentRequest()->get('q');
-        $manager = $this->container->get(DictionaryApiEntityManager::class);
-        $manager->setClass($manager->resolveEntityClass($dictionary, true));
-        $results = $manager->findValueBySearchQuery($searchQuery);
-        $responseContext = ['results' => $results];
-
-        return new JsonResponse($responseContext);
+        return $this->getJsonResponse(
+            $this->dictionaryEntityDataProvider->getValuesBySearchQuery($dictionary, $request->get('q'))
+        );
     }
 
     /**
-     * Get dictionary values by keys
-     *
-     * @param string $dictionary - Class Name Entity that was configured as Dictionary
-     *
-     *
-     * @return JsonResponse
+     * Gets dictionary values by keys.
      */
     #[Route(path: '/{dictionary}/values', name: 'oro_dictionary_value')]
-    public function valuesAction($dictionary)
+    public function valuesAction(string $dictionary, Request $request): JsonResponse
     {
-        $keys = $this->container->get('request_stack')->getCurrentRequest()->get('keys');
-        $manager = $this->container->get(DictionaryApiEntityManager::class);
-        $manager->setClass($manager->resolveEntityClass($dictionary, true));
-        $result = $manager->findValueByPrimaryKey($keys);
-        $responseContext = ['results' => $result];
+        $keys = $request->get('keys');
 
-        return new JsonResponse($responseContext);
+        return $this->getJsonResponse(
+            $keys ? $this->dictionaryEntityDataProvider->getValuesByIds($dictionary, $keys) : []
+        );
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public static function getSubscribedServices(): array
+    private function getJsonResponse(array $data): JsonResponse
     {
-        return array_merge(
-            parent::getSubscribedServices(),
-            [
-                DictionaryApiEntityManager::class,
-            ]
-        );
+        return new JsonResponse(['results' => $data]);
     }
 }