+++ src/Oro/Bundle/EntityExtendBundle/Form/Type/MultipleAssociationChoiceType.php
@@ -25,7 +25,7 @@ class MultipleAssociationChoiceType extends AbstractAssociationType
     {
         parent::buildForm($builder, $options);
 
-        $builder->addEventListener(FormEvents::SUBMIT, array($this, 'submit'));
+        $builder->addEventListener(FormEvents::SUBMIT, [$this, 'submit']);
     }
 
     /**
@@ -33,10 +33,10 @@ class MultipleAssociationChoiceType extends AbstractAssociationType
      */
     public function submit(FormEvent $event)
     {
-        $form    = $event->getForm();
+        $form = $event->getForm();
         $options = $form->getConfig()->getOptions();
         /** @var ConfigIdInterface $configId */
-        $configId  = $options['config_id'];
+        $configId = $options['config_id'];
         $className = $configId->getClassName();
 
         if (empty($className)) {
@@ -44,7 +44,7 @@ class MultipleAssociationChoiceType extends AbstractAssociationType
         }
 
         $immutable = $this->typeHelper->getImmutable($configId->getScope(), $className);
-        if (is_array($immutable) && !empty($immutable)) {
+        if (\is_array($immutable) && !empty($immutable)) {
             // set new values, but keep existing immutable values
             $existingValues = $this->configManager->getConfig($configId)->get($form->getName());
             if ($existingValues === null) {
@@ -66,11 +66,11 @@ class MultipleAssociationChoiceType extends AbstractAssociationType
         $resolver->setDefaults(
             [
                 'placeholder' => false,
-                'choices'     => function (Options $options) {
+                'choices' => function (Options $options) {
                     return $this->getChoices($options['association_class']);
                 },
-                'multiple'    => true,
-                'expanded'    => true,
+                'multiple' => true,
+                'expanded' => true,
                 'schema_update_required' => function ($newVal, $oldVal) {
                     if (!is_array($oldVal)) {
                         $oldVal = (array)$oldVal;
@@ -91,10 +91,9 @@ class MultipleAssociationChoiceType extends AbstractAssociationType
     public function finishView(FormView $view, FormInterface $form, array $options)
     {
         $disabledValues = $this->getReadOnlyValues($options);
-        /** @var FormView $choiceView */
         foreach ($view->children as $choiceView) {
             if ((isset($view->vars['disabled']) && $view->vars['disabled'])
-                || (!empty($disabledValues) && in_array($choiceView->vars['value'], $disabledValues))
+                || (!empty($disabledValues) && \in_array($choiceView->vars['value'], $disabledValues, true))
             ) {
                 $choiceView->vars['disabled'] = true;
             }
@@ -125,16 +124,11 @@ class MultipleAssociationChoiceType extends AbstractAssociationType
         return SymfonyChoiceType::class;
     }
 
-    /**
-     * @param string $groupName
-     *
-     * @return array
-     */
-    protected function getChoices($groupName)
+    private function getChoices(string $groupName): array
     {
-        $choices              = [];
+        $choices = [];
         $entityConfigProvider = $this->configManager->getProvider('entity');
-        $owningSideEntities   = $this->typeHelper->getOwningSideEntities($groupName);
+        $owningSideEntities = $this->associationTypeHelper->getOwningSideEntities($groupName);
         foreach ($owningSideEntities as $className) {
             $choices[$entityConfigProvider->getConfig($className)->get('plural_label')] = $className;
         }
@@ -142,22 +136,15 @@ class MultipleAssociationChoiceType extends AbstractAssociationType
         return $choices;
     }
 
-    /**
-     * Gets the list of values which state cannot be changed
-     *
-     * @param array $options
-     *
-     * @return string[]
-     */
-    protected function getReadOnlyValues(array $options)
+    private function getReadOnlyValues(array $options): array
     {
         /** @var EntityConfigId $configId */
-        $configId  = $options['config_id'];
+        $configId = $options['config_id'];
         $className = $configId->getClassName();
 
         if (!empty($className)) {
             $immutable = $this->typeHelper->getImmutable($configId->getScope(), $className);
-            if (is_array($immutable)) {
+            if (\is_array($immutable)) {
                 return $immutable;
             }
         }