+++ src/Oro/Bundle/QueryDesignerBundle/QueryDesigner/ConfigurationProvider.php
@@ -18,8 +18,7 @@ class ConfigurationProvider extends PhpArrayConfigProvider
     private const CONFIG_FILE = 'Resources/config/oro/query_designer.yml';
     private const APP_CONFIG_PATH = '../config/oro/query_designer';
 
-    /** @var Configuration */
-    private $configuration;
+    private Configuration $configuration;
 
     public function __construct(
         string $cacheFile,
@@ -39,16 +38,16 @@ class ConfigurationProvider extends PhpArrayConfigProvider
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
      */
-    protected function doLoadConfig(ResourcesContainerInterface $resourcesContainer)
+    protected function doLoadConfig(ResourcesContainerInterface $resourcesContainer): array
     {
         $configs = [];
         $configLoader = new CumulativeConfigLoader(
             'oro_query_designer',
             [
                 new YamlCumulativeFileLoader(self::CONFIG_FILE),
-                new FolderYamlCumulativeFileLoader($this->getAppConfigPath()),
+                new FolderYamlCumulativeFileLoader(self::APP_CONFIG_PATH),
             ]
         );
         $resources = $configLoader->load($resourcesContainer);
@@ -67,9 +66,6 @@ class ConfigurationProvider extends PhpArrayConfigProvider
         );
     }
 
-    /**
-     * Updates the label for all functions for the specified group type.
-     */
     private function updateLabelsOfFunctions(array &$config, string $groupType, string $vendor): void
     {
         if (isset($config[$groupType])) {
@@ -84,15 +80,6 @@ class ConfigurationProvider extends PhpArrayConfigProvider
         }
     }
 
-    /**
-     * Updates the label for the given function.
-     *
-     * @param array  $func
-     * @param string $labelType The type of label. Can be 'name' or 'hint'
-     * @param string $vendor
-     * @param string $groupType
-     * @param string $groupName
-     */
     private function updateFunctionLabel(
         array &$func,
         string $labelType,
@@ -114,10 +101,5 @@ class ConfigurationProvider extends PhpArrayConfigProvider
             // this function should use a label of overridden function
             $func[$labelName] = '';
         }
-    }
-
-    protected function getAppConfigPath(): string
-    {
-        return self::APP_CONFIG_PATH;
     }
 }